avoid dependencies on std::num
The `num` module has been stripped down on nightly, unfortunately. Apparently,
the `num` crate is the way to go now. But that crate is heavy and bloated, I
don't want it. This is not a big problem here, because it was just used as a
particular example to test, so now I use a different example to test.